[ARM cortex-m/user] Add support for cycle counter
To make up for the lack of a cycle count concept in user mode,
we add an instruction counter. Unlike earlier ARM processors, Cortex-M
memory map the performance counters. To enable the counters in user-mode
we reuse the original cp15_ccnt instruction here.

The original Cortex-M memory mapped cycle counters:

    void enable_cycle_counter(void) {
        DWT_CTRL |= DWT_CTRL_CYCCNTENA;
    }

    uint32_t __attribute__((noinline)) read_cycle_counter(void) {
        return __arm_mrc(15,0, 9, 13, 0); // ncp, opc1, crn, crm, opc2
        return DWT_CYCCNT;
    }